feature(add delete guard): Added delete guard to handleDelete functions from list services

+++ zafirasolidaire-website/src/services/event_services/EventList.jsx
@@ -32,6 +32,7 @@ const EventList = () => {
     }
 
     const handleDelete = async (id) => {
+        if (!window.confirm("Êtes-vous sûr(e) de vouloir supprimer cet évènement?")) return;
         try {
             await deleteEventById(id);
             setEvents(events.filter(e => e.id !== id));
